# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API системы управления заявками на полеты с JWT аутентификацией",
        "title": "Flight Request API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/flyRequests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список заявок с фильтрацией. Пользователи видят только свои заявки, модераторы - все.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Получить список заявок на полет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате формирования после",
                        "name": "formedAfter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате формирования до",
                        "name": "formedBefore",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую заявку на полет для текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Создать новую заявку на полет",
                "responses": {
                    "201": {
                        "description": "Созданная заявка",
                        "schema": {
                            "$ref": "#/definitions/ds.FlyRequest"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/current": {
            "get": {
                "description": "Возвращает ID текущей черновой заявки и количество румбов в ней",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Получить текущую заявку пользователя",
                "responses": {
                    "200": {
                        "description": "Информация о текущей заявке",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/rumbs/{flyRequestID}/{rumbID}": {
            "put": {
                "description": "Обновляет расстояние и скорость ветра для румба в заявке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequestRumbs"
                ],
                "summary": "Обновить параметры румба в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "flyRequestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "rumbID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет связь между румбом и заявкой на полет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequestRumbs"
                ],
                "summary": "Удалить румб из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "flyRequestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "rumbID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/{id}": {
            "get": {
                "description": "Возвращает полную информацию о заявке на полет включая связанные румбы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Получить детальную информацию о заявке на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детальная информация о заявке",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID заявки",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию о заявке на полет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Обновить заявку на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления заявки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.FlyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная заявка",
                        "schema": {
                            "$ref": "#/definitions/ds.FlyRequest"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет заявку на полет по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Удалить заявку на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Перенаправление на главную страницу"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/{id}/calculatedBy": {
            "put": {
                "description": "Обновляет информацию о том, кем был выполнен расчет заявки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Обновить поле CalculatedBy заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная информация",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/fly_calculation/{id}": {
            "get": {
                "description": "Отображает HTML страницу с детальной информацией о заявке на полет",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Frontend"
                ],
                "summary": "Получить страницу заявки на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML страница заявки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ID заявки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}